requirements.txt
fastapi[all]==0.75.1
pydantic==1.9.1
uvicorn==0.15.0
gunicorn==20.1.0
python-dotenv==0.20.0
confluent-kafka==1.8.2
starlette==0.17.1